refactor: extract pure-Node services to core/ (Stage A moves 1+2)
Move 1 — extract pure services (~9 of 17 files):
- 9 files moved electron/ → core/: browser-history-service,
  claude-presence, constants, nav-pins-service, pins-service,
  resolve-claude, session-state-service, skills-scanner, socket-server
- 5 importers updated (electron/main, install-service,
  browser-manager, pty-manager, scripts/postinstall)
- core/socket-server.ts cross-imports to electron-only modules
  (cdp-bridge, browser-manager, files-service) updated to ../electron/
- tsconfig.node.json includes core/**/*

Move 2 — EventSink for surface-coupled trio (partial):
- New core/event-sink.ts: host-agnostic event channel interface
- pty-manager moved to core/, refactored: WebContents → EventSink,
  app.getVersion() → constructor arg
- update-checker moved to core/, refactored: app.getVersion() →
  constructor arg, electron import dropped entirely
- electron/main.ts wraps webContents.send in a one-line EventSink
  adapter and passes app.getVersion() to constructors
- DEFERRED: files-service. Has more Electron coupling than the
  pre-refactor analysis assumed (also uses shell.openPath /
  showItemInFolder / trashItem); needs a SystemShell abstraction
  beyond the 0.5-day Move 2 budget.

Move 4 — formalize HostApi: no-op. The contract was already first-
class. shared/types.ts:825-1338 declares ElectronAPI with 21 sub-
interfaces, and electron/preload.ts:35 already does
`const api: ElectronAPI = {...}`. Cosmetic rename (Electron* → Host*)
deferred until extension lands.

Net effect:
- electron/ down from 17 files to 7 (genuinely Electron-coupled:
  main, preload, browser-manager, cdp-bridge, files-service,
  install-service, auto-updater)
- core/ at 12 files (services + EventSink + constants)
- npm run typecheck and npm run build both clean
- No behavior change to the running Electron app

This is the no-regrets refactor identified in
docs/research/duo-as-chrome-extension/refactor-analysis.md — pays off
whether or not the Chrome extension shape is ever walked, by giving
services a testable home outside Electron and clarifying what's truly
Electron-coupled.
