+++ core/browser-history-service.ts
