+++ core/claude-presence.ts
