+++ core/session-state-service.ts
