+++ core/skills-scanner.ts
